requirements.txt
requests
tenacity
openai
streamlit
pandas
python-dotenv
